requirements.txt
asyncpg==0.27.0
fastapi==0.94.1
gunicorn==20.1.0
httpx==0.23.3
uvicorn==0.21.1
SQLAlchemy==2.0.15
psycopg2==2.9.6
python-dotenv==1.0.0
alembic==1.11.1
pytest-asyncio-0.21.0
